// src/api/models/General.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const General = {
    NUMBER_0: 0,
    NUMBER_1: 1,
    NUMBER_2: 2,
    NUMBER_3: 3,
    NUMBER_4: 4,
    NUMBER_5: 5,
    NUMBER_6: 6,
    NUMBER_7: 7,
    NUMBER_8: 8,
    NUMBER_9: 9,
    NUMBER_10: 10,
    NUMBER_11: 11,
    NUMBER_MINUS_1: -1
} as const;
export type General = typeof General[keyof typeof General];


export function instanceOfGeneral(value: any): boolean {
    for (const key in General) {
        if (Object.prototype.hasOwnProperty.call(General, key)) {
            if (General[key as keyof typeof General] === value) {
                return true;
            }
        }
    }
    return false;
}

export function GeneralFromJSON(json: any): General {
    return GeneralFromJSONTyped(json, false);
}

export function GeneralFromJSONTyped(json: any, ignoreDiscriminator: boolean): General {
    return json as General;
}

export function GeneralToJSON(value?: General | null): any {
    return value as any;
}

export function GeneralToJSONTyped(value: any, ignoreDiscriminator: boolean): General {
    return value as General;
}