// src/api/models/WinLoss.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface WinLoss
 */
export interface WinLoss {
    /**
     * 
     * @type {number}
     * @memberof WinLoss
     */
    wins: number;
    /**
     * 
     * @type {number}
     * @memberof WinLoss
     */
    losses: number;
}

/**
 * Check if a given object implements the WinLoss interface.
 */
export function instanceOfWinLoss(value: object): value is WinLoss {
    if (!('wins' in value) || value['wins'] === undefined) return false;
    if (!('losses' in value) || value['losses'] === undefined) return false;
    return true;
}

export function WinLossFromJSON(json: any): WinLoss {
    return WinLossFromJSONTyped(json, false);
}

export function WinLossFromJSONTyped(json: any, ignoreDiscriminator: boolean): WinLoss {
    if (json == null) {
        return json;
    }
    return {
        
        'wins': json['wins'],
        'losses': json['losses'],
    };
}

export function WinLossToJSON(json: any): WinLoss {
    return WinLossToJSONTyped(json, false);
}

export function WinLossToJSONTyped(value?: WinLoss | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'wins': value['wins'],
        'losses': value['losses'],
    };
}